requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.43.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==44.0.2
Django==5.1.7
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-ckeditor==6.7.2
django-cors-headers==4.7.0
django-js-asset==3.1.2
django-oauth-toolkit==3.0.1
django-timezone-field==7.1
djangorestframework==3.15.2
dnspython==2.7.0
dotenv==0.9.9
drf-yasg==1.21.10
email_validator==2.2.0
idna==3.10
inflection==0.5.1
jwcrypto==1.5.6
kombu==5.5.0
msgpack==1.1.0
oauthlib==3.2.2
packaging==24.2
prompt_toolkit==3.0.50
pycparser==2.22
PyMySQL==1.1.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
